+++ app/Http/Controllers/BusinessUserController.php
@@ -3,21 +3,32 @@
 namespace Siac\Http\Controllers;
 
 use Illuminate\Http\Request;
-use Yajra\Datatables\Datatables;
 use Siac\BusinessUser;
 use Siac\Company;
 use Siac\Http\Requests\BusinessUserRequest;
+
+use Yajra\Datatables\Datatables;
+use Yajra\DataTables\Services\DataTable;
 
 class BusinessUserController extends Controller
 {
     public function index(){
 
-        
-        $businessusers = BusinessUser::OrderBy('id','asc')->with(['relationCompany' => function($query){
-                        $query->select('id','nameCompany');
-        }])->get();
-         
         return view('businessusers.index', compact('businessusers'));
+    }
+
+    public function getUserCompany()
+    {
+        
+        $businessusers = BusinessUser::with(['companies' => function($query){
+                                                $query->select('id','nameCompany');
+                                            }]);
+        
+        return Datatables::of($businessusers)->addColumn('action', function ($user) {
+            return '<a href="http://www.sipcom.cl/siac/businessusers/'.$user->id.'" class="btn btn-sm btn-success"><i class="material-icons">pageview</i></a>
+                    <a href="http://www.sipcom.cl/siac/businessusers/'.$user->id.'/edit" class="btn btn-sm btn-info"><i class="material-icons">border_color</i></a>
+                    <a href="#" onclick="destroyUserCompany('.$user->id.')" class="btn btn-sm btn-warning"><i class="material-icons">delete_forever</i></a>';
+        })->make(true);
     }
 
     public function create(){
@@ -71,13 +82,17 @@ class BusinessUserController extends Controller
         
     }
 
-    public function destroy($id){
+    public function destroyUserCompany(request $request){
 
-        $businessuser = BusinessUser::find($id);
+        if($request->ajax())
+        {
+            $businessuser = BusinessUser::find($request->id);
 
-        $businessuser->delete();
+            $businessuser->delete();
 
-        return back()->with('info', 'Se ha eliminado con éxito');
+            return 'Se ha eliminado el usuario de empresa';
+        }
+        
     }
 
     public function getTasks()
